pyproject.toml
[project]
name = "lang-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=1.1.0",
    "langchain-google-genai>=3.1.0",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
]
